rename page-- to just {name}
